Moving join us to footer
