+++ components/LandingPage/index.ts
@@ -2,5 +2,4 @@ export { Hero } from './Hero';
 export { Guide } from './Guide';
 export { About } from './About';
 export { Thanks } from './Thanks';
-export { JoinUs } from './JoinUs';
 export { Benefit, BenefitsContainer } from './Benefit';
