+++ components/Layouts/SharedLayout.tsx
@@ -3,6 +3,9 @@ import React from 'react';
 import { AuthModal } from '@/components/AuthModal';
 import { Topbar } from '../Topbar/Topbar';
 import Link from 'next/link';
+import { Typography } from '../Typography';
+import { motion } from 'framer-motion';
+import { getSocialLinks } from '../Topbar/data';
 
 interface IProps {
   children: React.ReactNode;
@@ -18,7 +21,7 @@ export const SharedLayout: React.FC<IProps> = ({
   hasContainer,
 }) => {
   const year = new Date().getFullYear();
-
+  const socialLinks = getSocialLinks();
   return (
     <>
       <Head>
@@ -32,6 +35,30 @@ export const SharedLayout: React.FC<IProps> = ({
       <AuthModal title={'Continue with your social accounts'} />
       {!hideFooter && (
         <footer className="z-100 sticky top-[100%] border-t px-5 py-10 text-center dark:border-gray-800">
+          <div className="flex flex-col items-center justify-center rounded-full p-5">
+            <Typography
+              as="h2"
+              fontSize="2xl"
+              fontWeight="bold"
+              className="mb-10 uppercase"
+            >
+              join us
+            </Typography>
+            <div className="flex flex-row items-center justify-center space-x-16">
+              {socialLinks.map(({ Icon, title, anchorTagProps, url }) => (
+                <motion.a
+                  whileHover={{ scale: 1.2 }}
+                  key={title}
+                  {...anchorTagProps}
+                  href={url}
+                  className="flex items-center justify-center text-5xl text-slate-600 hover:text-slate-900 dark:text-slate-400 dark:hover:text-slate-100"
+                >
+                  {<Icon />}
+                </motion.a>
+              ))}
+            </div>
+          </div>
+          <br />
           <ul className="my-4 flex flex-col items-center justify-center gap-2 md:flex-row md:gap-8">
             <li>
               <Link href="/privacy-policy">Privacy Policy</Link>
